Added new unit tests on DayUnit and deleted on RunSpecifier
