+++ UnitTests/Fluent/1 - Run/RunSpecifierTests.cs
@@ -148,74 +148,5 @@
             // Assert
             Assert.AreEqual(null, calculated);
         }
-		[TestMethod]
-		public void EveryWeekday()
-		{
-			// Arrange
-			var sunday = new DateTime(2018, 02, 18);
-			var saturday = new DateTime(2018, 02, 17);
-
-			var monday = new DateTime(2018, 02, 19);
-
-			var calculator = new TimeCalculator();
-			var run = new RunSpecifier(calculator);
-
-			// Act
-			run.EveryWeekday();
-			var calculated = calculator.Calculate(sunday);
-
-			// Assert
-			var expected = new DateTime(2018, 02, 19);
-			Assert.AreEqual(expected, calculated);
-
-			// Act
-			run.EveryWeekday();
-			calculated = calculator.Calculate(saturday);
-
-			// Assert
-			Assert.AreEqual(expected, calculated);
-
-			// Act
-			run.EveryWeekday();
-			calculated = calculator.Calculate(monday);
-
-			// Assert
-			Assert.AreEqual(expected, calculated);
-		}
-
-		[TestMethod]
-		public void EveryWeekend()
-		{
-			// Arrange
-			var tuesday = new DateTime(2018, 02, 13);
-			var saturday = new DateTime(2018, 02, 17);
-			var sunday = new DateTime(2018, 02, 18);
-
-			var calculator = new TimeCalculator();
-			var run = new RunSpecifier(calculator);
-
-			// Act
-			run.EveryWeekend();
-			var calculated = calculator.Calculate(tuesday);
-
-			// Assert
-			var expected = new DateTime(2018, 02, 17);
-			Assert.AreEqual(expected, calculated);
-
-			// Act
-			run.EveryWeekend();
-			calculated = calculator.Calculate(saturday);
-
-			// Assert
-			Assert.AreEqual(expected, calculated);
-
-			// Act
-			run.EveryWeekend();
-			calculated = calculator.Calculate(sunday);
-
-			// Assert
-			expected = new DateTime(2018, 02, 18);
-			Assert.AreEqual(expected, calculated);
-		}
 	}
 }
